pyproject.toml
[project]
name = "ai-style-transfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fiftyone>=1.10.0",
    "huggingface-hub>=1.1.4",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]
